+++ test/controllers/location.js
@@ -4,32 +4,13 @@ process.env.NODE_ENV = 'test';
 const request = require('supertest');
 const helper = require('../helper');
 
-const https = require('https')
-
-async function getPublicIp() { 
-  const url = 'https://api64.ipify.org?format=json'
-
-  return new Promise((resolve, reject) => {
-    var data = '';
-    https.get(url, res => {
-      res.on('data', chunk => { data += chunk })
-      res.on('end', () => {
-        const parse = JSON.parse(data);
-        if(parse.ip)
-          return resolve(parse.ip);
-        return reject();
-      })
-    })
-  });
-}
-
 describe("Find location of unlogged user", async function () {
   let server;
   let publicIp;
 
   before(async function () {
     try {
-      publicIp = await getPublicIp();
+      publicIp = await helper.getPublicIp();
     } catch (e) {
       publicIp = '191.213.233.162';
     }
